[fingerprint] Convert cli to typescript
