+++ packages/@expo/fingerprint/cli/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require('expo-module-scripts/jest-preset-cli');
